openssh_fixture.c: fix compiler warnings
```
tests/openssh_fixture.c:116:..: warning C4267: 'function': conversion from 'size_t' to 'int', possible loss of data
tests/openssh_fixture.c:116:42: warning: conversion from 'long long unsigned int' to 'int' may change value [-Wconversion]
tests/openssh_fixture.c:116:42: warning: conversion from 'long unsigned int' to 'int' may change value [-Wconversion]
tests/openssh_fixture.c:116:42: warning: implicit conversion loses integer precision: 'unsigned long long' to 'int' [-Wshorten-64-to-32]
tests/openssh_fixture.c:116:42: warning: implicit conversion loses integer precision: 'unsigned long' to 'int' [-Wshorten-64-to-32]
```

+++ tests/openssh_fixture.c
@@ -113,7 +113,7 @@ static int run_command_varg(char **output, const char *command, va_list args)
     buf[0] = 0;
     buf_len = 0;
     while(buf_len < (sizeof(buf) - 1) &&
-        fgets(&buf[buf_len], sizeof(buf) - buf_len, pipe) != NULL) {
+        fgets(&buf[buf_len], (int)(sizeof(buf) - buf_len), pipe) != NULL) {
         buf_len = strlen(buf);
     }
